+++ shipcat_cli/tests/helm.rs
@@ -2,15 +2,18 @@ mod common;
 use crate::common::setup;
 use shipcat_definitions::{Config, ConfigType};
 use shipcat::helm;
+use shipcat::Result;
 
 #[test]
 #[ignore] // This test requires helm cli - not on circle
-fn helm_template() {
+fn helm_template() -> Result<()> {
     setup();
-    let (conf, reg) = Config::new(ConfigType::Base, "dev-uk").unwrap();
-    let mock = true;
-    let res = helm::template("fake-ask", &reg, &conf, None, mock, None).unwrap();
+    let (conf, reg) = Config::new(ConfigType::Base, "dev-uk")?;
+    let mf = shipcat_filebacked::load_manifest("fake-storage", &conf, &reg)?.stub(&reg)?;
+
+    let res = helm::template(&mf, None)?;
 
     // verify we have deferred to helm for templating
     assert!(res.contains("image: \"quay.io/babylonhealth/fake-ask:1.6.0\""));
+    Ok(())
 }